make testing

+++ src/main/java/com/zadanie/zadanie2/service/BookingService.java
@@ -1,5 +1,6 @@
 package com.zadanie.zadanie2.service;
 
+import com.zadanie.zadanie2.exceptions.RecordNotFoundException;
 import com.zadanie.zadanie2.model.Booking;
 import com.zadanie.zadanie2.model.Room;
 import com.zadanie.zadanie2.repository.BookingRepository;
@@ -18,7 +19,11 @@ public class BookingService {
     private RoomRepository roomRepository;
 
     public List<Booking> findAllBooking(){
-        return bookingRepository.findAll();
+        List<Booking> bookings = bookingRepository.findAll();
+        if(bookings.isEmpty()){
+            throw new RecordNotFoundException("You are too poor to have reservations");
+        }
+        return bookings;
     }
     public Booking saveBooking(Booking booking){
         return bookingRepository.save(booking);
@@ -29,17 +34,36 @@ public class BookingService {
     }
 
     public List<Booking> findByRoomId(Long roomId){
-        Room room = roomRepository.findRoomById(roomId);
-
-        return bookingRepository.findByRoom(room);
+        List<Booking> bookings = bookingRepository.findById(roomId);
+        if(bookings.isEmpty()){
+            throw new RecordNotFoundException("Don't found any bookings on this room");
+        }
+        return bookings;
     }
 
     public List<Booking> findBookingByGuestName(String guestName){
-        return bookingRepository.findByGuestName(guestName);
+        List<Booking> bookings = bookingRepository.findByGuestName(guestName);
+        if(bookings.isEmpty()){
+            throw new RecordNotFoundException("Don't found booking for this guestName");
+        }
+        return bookings;
+    }
+
+    public List<Booking> findByBookingId(Long id){
+        List<Booking> bookings = bookingRepository.findById(id);
+        if(bookings.isEmpty()){
+            throw new RecordNotFoundException("This booking doesn't exist");
+        }
+        return bookings;
     }
 
     public void dropBooking(Long id){
+        List<Booking> bookings = bookingRepository.findById(id);
+        if(bookings.isEmpty()){
+            throw new RecordNotFoundException("This booking doesn't exist");
+        }
         bookingRepository.delete(id);
+
     }
 
 }
