+++ src/main/java/com/zadanie/zadanie2/service/RoomService.java
@@ -1,5 +1,9 @@
 package com.zadanie.zadanie2.service;
 
+import com.zadanie.zadanie2.exceptions.RecordNotFoundException;
+import com.zadanie.zadanie2.helpers.ApiResponse;
+import com.zadanie.zadanie2.helpers.ApiResponse.Status;
+import com.zadanie.zadanie2.helpers.ApiResponse.ApiError;
 import com.zadanie.zadanie2.model.Room;
 import com.zadanie.zadanie2.repository.RoomRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,18 +17,30 @@ public class RoomService {
     private RoomRepository roomRepository;
 
     public List<Room> findAll(){
-        return roomRepository.findAll();
+        List<Room> rooms = roomRepository.findAll();
+        if(rooms.isEmpty()){
+            throw new RecordNotFoundException("No any rooms");
+        }
+        return rooms;
     }
 
     public Room insertRoom(Room room){
         return roomRepository.save(room);
     }
 
-    public Room findRoomById(Long id){
-        return roomRepository.findRoomById(id);
+    public List<Room> findRoomById(Long id){
+        List<Room> rooms = roomRepository.findRoomById(id);
+        if(rooms == null){
+            throw new RecordNotFoundException("No room with ID " + id);
+        }
+        return rooms;
     }
 
     public void dropRoom(Long id){
+        List<Room> rooms = roomRepository.findRoomById(id);
+       if(rooms.isEmpty()){
+           throw new RecordNotFoundException("I can't find record to remove");
+       }
        roomRepository.delete(id);
     }
     public Room updateData(Room room){
